pyproject.toml
[project]
name = "handsoff-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.5.5",
    "graphviz>=0.20.3",
    "litellm>=1.67.5",
    "openai-agents>=0.0.13",
    "tavily-python>=0.7.2",
]
